[snitch] :bug: Do not write back L/S post-incremented address if stall
This commit fixes a bug in the implementation of the Xpulpimg post-increment load/store instructions: such instructions
write back on rs1 the updated address; if the core stalled (for operands not ready or LSU not ready for a new request)
in the cycle during which a post-increment instruction was issued, the write-back mechanism updated anyway the new rs1
in the register file, modifying the address on the LSU interface on next cycle, while the request of the memory access
was still pending. The fix makes the rs1 write-back stall together with the core when a post-increment instruction is
fetched; rs1 is actually written back with the post-incremented address only in the cycle when the LSU actually accepts
the new request.